GH-90: Moved common auth functionality into 'common' module
