+++ common/auth/src/main/java/com/nucleonforge/axile/common/auth/Authorizer.java
@@ -1,0 +1,23 @@
+package com.nucleonforge.axile.common.auth;
+
+import com.nucleonforge.axile.common.auth.core.AuthorizationRequest;
+import com.nucleonforge.axile.common.auth.core.User;
+import com.nucleonforge.axile.common.auth.exception.AuthorizationException;
+
+/**
+ * SPI interface that is capable of authorizing the given {@link User} against an {@link AuthorizationRequest}.
+ *
+ * @since 16.07.25
+ * @author Mikhail Polivakha
+ */
+public interface Authorizer {
+
+    /**
+     * Authorizes the given {@link User} against the specified {@link AuthorizationRequest}.
+     *
+     * @param user the user to authorize
+     * @param authorizationRequest the request containing required authorities
+     * @throws AuthorizationException if access is denied
+     */
+    void authorize(User user, AuthorizationRequest authorizationRequest) throws AuthorizationException;
+}
