+++ common/auth/src/main/java/com/nucleonforge/axile/common/auth/exception/JwtParsingException.java
@@ -1,0 +1,14 @@
+package com.nucleonforge.axile.common.auth.exception;
+
+/**
+ * Indicates that an error occurred while parsing a JWT (JSON Web Token).
+ *
+ * @since 26.08.2025
+ * @author Nikita Kirillov
+ */
+public class JwtParsingException extends RuntimeException {
+
+    public JwtParsingException(String message) {
+        super(message);
+    }
+}
