+++ common/auth/src/test/java/com/nucleonforge/axile/common/auth/Main.java
@@ -1,0 +1,17 @@
+package com.nucleonforge.axile.common.auth;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+/**
+ * Minimal Spring Boot application used exclusively for testing this application.
+ *
+ * @since 17.07.2025
+ * @author Nikita Kirillov
+ */
+@SpringBootApplication
+public class Main {
+    public static void main(String[] args) {
+        SpringApplication.run(Main.class, args);
+    }
+}
