+++ common/auth/src/test/resources/application.yml
