+++ master/src/test/java/com/nucleonforge/axile/master/auth/spi/jwt/DefaultJwtEncoderServiceTest.java
@@ -27,7 +27,7 @@ import com.nucleonforge.axile.common.auth.core.DefaultRole;
 import com.nucleonforge.axile.common.auth.core.DefaultUser;
 import com.nucleonforge.axile.common.auth.core.Role;
 import com.nucleonforge.axile.common.auth.core.User;
-import com.nucleonforge.axile.common.auth.spi.jwt.JwtAlgorithm;
+import com.nucleonforge.axile.common.auth.jwt.JwtAlgorithm;
 import com.nucleonforge.axile.master.auth.spi.jwt.service.DefaultJwtEncoderService;
 import com.nucleonforge.axile.master.auth.spi.jwt.service.JwtEncoderService;
 import com.nucleonforge.axile.master.auth.spi.jwt.service.JwtTokenGenerationException;
