--- sbs/auth/src/main/java/com/nucleonforge/axile/auth/spi/jwt/exception/JwtParsingException.java
@@ -1,14 +1,0 @@
-package com.nucleonforge.axile.auth.spi.jwt.exception;
-
-/**
- * Indicates that an error occurred while parsing a JWT (JSON Web Token).
- *
- * @since 26.08.2025
- * @author Nikita Kirillov
- */
-public class JwtParsingException extends RuntimeException {
-
-    public JwtParsingException(String message) {
-        super(message);
-    }
-}
